feat:レイヤードアーキテクチャを用いたアーキテクチャの再設計完了
