+++ src/application/mod.rs
@@ -1,0 +1,1 @@
+pub mod utils;
