+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod application;
+pub mod domain;
+pub mod infra;
+pub mod presentation;
